requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.3
django-admin-honeypot-updated-2021==1.2.0
django-allauth==65.2.0
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-ipware==7.0.1
django-multiselectfield==0.1.13
django-resized==1.0.3
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-ipware==3.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
